+++ BprsLoanReport/BprsLoanReportService/src/org/simbiosis/report/loan/LoanPaymentExtPdf.java
@@ -33,7 +33,7 @@ public class LoanPaymentExtPdf extends ReportServlet {
 	String code;
 	String savingCode;
 	Double principal = 0D;
-	Double tenor = 0D;
+	Integer tenor = 0;
 
 	public LoanPaymentExtPdf() {
 		super("LoanPaymentExtPdf");
